backend/requirements.txt
fastapi
uvicorn
transformers
torch
pillow
python-multipart
moviepy==1.0.3
opencv-python-headless